+++ lib/Hakyll/Web/Pandoc.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE CPP #-}
 --------------------------------------------------------------------------------
 -- | Module exporting convenient pandoc bindings
 module Hakyll.Web.Pandoc
@@ -67,6 +68,17 @@ readPandocWith ropt item =
         MediaWiki          -> readMediaWiki ro
         OrgMode            -> readOrg ro
         Rst                -> readRST ro
+#if MIN_VERSION_pandoc(3,8,3)
+        AsciiDoc           -> readAsciiDoc ro
+#endif
+#if MIN_VERSION_pandoc(3,1,12)
+        Djot               -> readDjot ro
+#endif
+-- This preprocessing instruction can be dropped
+-- once the minimum supported GHC version is 8.10
+#if MIN_VERSION_pandoc(3,1,3)
+        Typst              -> readTypst ro
+#endif
         Textile            -> readTextile ro
         _                  -> error $
             "Hakyll.Web.readPandocWith: I don't know how to read a file of " ++
@@ -224,7 +236,14 @@ defaultHakyllWriterOptions = def
       writerExtensions = enableExtension Ext_smart pandocExtensions
     , -- We want to have hightlighting by default, to be compatible with earlier
       -- Hakyll releases
+#if MIN_VERSION_pandoc(3,8,0)
+      -- Starting with pandoc 3.8, the highlighting
+      -- system was overhauled to have more than just Skylighting
+      -- styles
+      writerHighlightMethod = Skylighting pygments
+#else
       writerHighlightStyle = Just pygments
+#endif
     , -- Do not word-wrap produced HTML, and do not undo any word-wrapping
       -- that's already present in the markup. This is how Pandoc operated
       -- prior to 2.17, but the behaviour was changed for consistency with