+++ lib/Hakyll/Web/Pandoc/FileType.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE CPP #-}
 --------------------------------------------------------------------------------
 -- | A module dealing with pandoc file extensions and associated file types
 module Hakyll.Web.Pandoc.FileType
@@ -33,6 +34,17 @@ data FileType
     | PlainText
     | Rst
     | Textile
+#if MIN_VERSION_pandoc(3,8,3)
+    | AsciiDoc
+#endif
+#if MIN_VERSION_pandoc(3,1,12)
+    | Djot
+#endif
+-- This preprocessing instruction can be dropped
+-- once the minimum supported GHC version is 8.10
+#if MIN_VERSION_pandoc(3,1,3)
+    | Typst
+#endif
     deriving (Eq, Ord, Show, Read)
 
 
@@ -66,6 +78,19 @@ fileType = uncurry fileType' . splitExtension
     fileType' _ ".text"      = PlainText
     fileType' _ ".textile"   = Textile
     fileType' _ ".txt"       = PlainText
+#if MIN_VERSION_pandoc(3,8,3)
+    fileType' _ ".asciidoc"  = AsciiDoc
+    fileType' _ ".adoc"      = AsciiDoc
+#endif
+#if MIN_VERSION_pandoc(3,1,12)
+    fileType' _ ".dj"        = Djot
+    fileType' _ ".djot"      = Djot
+#endif
+-- This preprocessing instruction can be dropped
+-- once the minimum supported GHC version is 8.10
+#if MIN_VERSION_pandoc(3,1,3)
+    fileType' _ ".typ"       = Typst
+#endif
     fileType' _ ".wiki"      = MediaWiki
     fileType' _ _            = Binary  -- Treat unknown files as binary